task_4/requirements.txt
apache-airflow
matplotlib
pandas
prophet
scikit-learn
yfinance